starting migrate and handle palette logic
